Update application version, improve error handling, Add game specific patches
The application version is updated to 3.5 and better error handling has been added throughout the code. A number of write operations are now within a try-catch block to mitigate the potential for crashes. Several changes are also made to improve how the software handles missing elements and cleans up after execution.